vet fixes
